+++ src/constants.rs
@@ -131,49 +131,3 @@ pub mod physical {
     /// Light speed (cm/s)
     pub const LIGHT_SPEED: f64 = 2.99792458e+10;
 }
-
-/// Mesh and geometry related constants
-pub mod mesh {
-    //!
-    //! The used mesh is made of cells (hexahedron), each divided in 12 sub-cells (tetrahedron).
-    //!
-    //! TODO: insert an image?
-    //!
-    use super::Tuple4;
-
-    /// Number of points per tetrahedron facet.
-    pub const N_POINTS_PER_FACET: usize = 3;
-    /// Number of facets of a cell facing outward i.e. constituting
-    /// a border with another cell.
-    pub const N_FACETS_OUT: usize = 24;
-    /// Number of points defining a cell.
-    pub const N_POINTS_INTERSEC: usize = 14;
-    /// Offsets of the intersection points of a cell.
-    pub const CORNER_OFFSET: [Tuple4; N_POINTS_INTERSEC] = [
-        (0, 0, 0, 0),
-        (1, 0, 0, 0),
-        (0, 1, 0, 0),
-        (1, 1, 0, 0),
-        (0, 0, 1, 0),
-        (1, 0, 1, 0),
-        (0, 1, 1, 0),
-        (1, 1, 1, 0),
-        (1, 0, 0, 1),
-        (0, 0, 0, 1),
-        (0, 1, 0, 2),
-        (0, 0, 0, 2),
-        (0, 0, 1, 3),
-        (0, 0, 0, 3),
-    ];
-    /// Number of faces defining a cell.
-    pub const N_FACES: usize = 6;
-    /// Offsets of the faces of a cell.
-    pub const FACE_OFFSET: [(i32, i32, i32); N_FACES] = [
-        (1, 0, 0),
-        (-1, 0, 0),
-        (0, 1, 0),
-        (0, -1, 0),
-        (0, 0, 1),
-        (0, 0, -1),
-    ];
-}
